requirements.txt
discord.py
pretty_help
dnspython
platform
Embed